tmp: move internal type to pkg/apis/componentconfig/types.go
