+++ staging/src/k8s.io/apiserver/pkg/admission/plugin/webhook/config/apis/webhookadmission/register.go
@@ -19,6 +19,7 @@ package webhookadmission
 import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/kubernetes/pkg/apis/componentconfig"
 )
 
 var (
@@ -45,7 +46,7 @@ func Resource(resource string) schema.GroupResource {
 func addKnownTypes(scheme *runtime.Scheme) error {
 	// TODO this will get cleaned up with the scheme types are fixed
 	scheme.AddKnownTypes(SchemeGroupVersion,
-		&EncryptionConfiguration{},
+		&componentconfig.EncryptionConfiguration{},
 		&WebhookAdmission{},
 	)
 	return nil
