+++ staging/src/k8s.io/apiserver/pkg/admission/plugin/webhook/config/apis/webhookadmission/types.go
@@ -27,71 +27,3 @@ type WebhookAdmission struct {
 	// KubeConfigFile is the path to the kubeconfig file.
 	KubeConfigFile string
 }
-
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// EncryptionConfiguration stores the complete configuration for encryption providers.
-type EncryptionConfiguration struct {
-	metav1.TypeMeta
-	// resources is a list containing resources, and their corresponding encryption providers.
-	Resources []ResourceConfiguration
-}
-
-// ResourceConfiguration stores per resource configuration.
-type ResourceConfiguration struct {
-	// resources is a list of kubernetes resources which have to be encrypted.
-	Resources []string
-	// providers is a list of transformers to be used for reading and writing the resources to disk.
-	// eg: aesgcm, aescbc, secretbox, identity.
-	Providers []ProviderConfiguration
-}
-
-// ProviderConfiguration stores the provided configuration for an encryption provider.
-type ProviderConfiguration struct {
-	// aesgcm is the configuration for the AES-GCM transformer.
-	AESGCM *AESConfiguration
-	// aescbc is the configuration for the AES-CBC transformer.
-	AESCBC *AESConfiguration
-	// secretbox is the configuration for the Secretbox based transformer.
-	Secretbox *SecretboxConfiguration
-	// identity is the (empty) configuration for the identity transformer.
-	Identity *IdentityConfiguration
-	// kms contains the name, cache size and path to configuration file for a KMS based envelope transformer.
-	KMS *KMSConfiguration
-}
-
-// AESConfiguration contains the API configuration for an AES transformer.
-type AESConfiguration struct {
-	// keys is a list of keys to be used for creating the AES transformer.
-	// Each key has to be 32 bytes long for AES-CBC and 16, 24 or 32 bytes for AES-GCM.
-	Keys []Key
-}
-
-// SecretboxConfiguration contains the API configuration for an Secretbox transformer.
-type SecretboxConfiguration struct {
-	// keys is a list of keys to be used for creating the Secretbox transformer.
-	// Each key has to be 32 bytes long.
-	Keys []Key
-}
-
-// Key contains name and secret of the provided key for a transformer.
-type Key struct {
-	// name is the name of the key to be used while storing data to disk.
-	Name string
-	// secret is the actual key, encoded in base64.
-	Secret string
-}
-
-// IdentityConfiguration is an empty struct to allow identity transformer in provider configuration.
-type IdentityConfiguration struct{}
-
-// KMSConfiguration contains the name, cache size and path to configuration file for a KMS based envelope transformer.
-type KMSConfiguration struct {
-	// name is the name of the KMS plugin to be used.
-	Name string
-	// cacheSize is the maximum number of secrets which are cached in memory. The default value is 1000.
-	// +optional
-	CacheSize int
-	// endpoint is the gRPC server listening address, for example "unix:///var/run/kms-provider.sock".
-	Endpoint string
-}
